Move UnknownDiscoverItemAlert to a separate file
